Extract contacts-messaging composition hook from useAppShellComposition
